+++ fawltydeps/traverse_project.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import AbstractSet, Iterator, Optional, Set, Tuple, Type, Union
 
 from fawltydeps.dir_traversal import DirectoryTraversal
-from fawltydeps.extract_declared_dependencies import validate_deps_source
+from fawltydeps.extract_deps import validate_deps_source
 from fawltydeps.extract_imports import validate_code_source
 from fawltydeps.gitignore_parser import RuleError as ExcludeRuleError
 from fawltydeps.packages import validate_pyenv_source
